COMMERCE-913 Refactor package name
